En esta rama, tuve que modificar 3 clases, 1) UsuarioService: Se aniade una condicion IF en eliminarUsuario para validar si responde correctamente. En obtenerPorId deje solo null para una mejor flexibilidad en el modelo de HATEO. 2) UsuarioModelAssembler: Se crea para encapsular la logica del modelo Usuario con HATEOAS. Esto nos ayuda a crear enlaces quue indican acciones o rutas relacionadas. Los enlaces de EntityModel quedaron como: self, usuarios, actualizar y eliminar. El metodo wrapError fue creado para encapsular las respuestas de error, toCollection: funciona para listar multiples usuarios con enlaces. Todo esto nos ayuda a evitar duplicacion de logica en el controlador, incluso con los mensajes de error en la navegacion. 3) UsuarioControlador: Finalizando con esta clase, se elimina los errores 500 ya que estos son innecesarios, se mejora la gramatica de respuestas paraa que sea mas intuitivo con el cliente. Ademas, es aqui tambien en los endpoints en donde se agrega HAL Hypertext Application Language que viene siendo el conjunto de enlaces LINKS para una mejor navegacion y una mejor documentacion legible para el cliente, ademas de sentencias Try para manejo de errores

+++ usUsuario/src/main/java/com/petsocity/petsocity/service/UsuarioService.java
@@ -23,8 +23,8 @@ public class UsuarioService {
         return usuarioRepository.findAll();
     }
 
-    public Optional<Usuario> obtenerPorIdUsuario(Long id) {
-        return usuarioRepository.findById(id);
+    public Usuario obtenerPorIdUsuario(Long id) {
+        return usuarioRepository.findById(id).orElse(null);
     }
 
     public Usuario crearUsuario(Usuario usuario) {
@@ -79,7 +79,13 @@ public class UsuarioService {
         }).orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
     }
 
-    public void eliminarUsuario(Long id) {
-        usuarioRepository.deleteById(id);
+    public boolean eliminarUsuario(Long id) {
+    Optional<Usuario> usuario = usuarioRepository.findById(id);
+        if (usuario.isPresent()) {
+            usuarioRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 }
